pyproject.toml
[tool.poetry]
name = "rbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.4.0"
pytesseract = "^0.3.13"
pillow = "^11.1.0"
fpdf2 = "^2.8.2"
httpx = "^0.28.1"
python-dotenv = "^1.0.1"
pyphen = "^0.17.2"
google-generativeai = "^0.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
